Swift

+++ backend/deploy.sh
@@ -1,4 +1,9 @@
 git reset --hard
 git pull
+rm -f Package.resolved
 swift build
-cp -r .build/debug/CanaryBackend release/canary
+if [ -f ".build/release/CanaryBackend" ];then
+    cp -r .build/release/CanaryBackend release/canary
+  else  
+    cp -r .build/debug/CanaryBackend release/canary
+fi
